pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.0",
    "portia-sdk-python[all]>=0.7.2",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
]
